chore: moved inner class of MockExtension into own files
