+++ JotunnLib/Extensions/MockExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using Jotunn.Managers;
 using Jotunn.Utils;
@@ -73,124 +72,6 @@ namespace Jotunn
                 {
                     tf.gameObject.FixReferences(true);
                 }*/
-            }
-        }
-
-        private class ClassMember
-        {
-            public readonly List<MemberBase> members = new List<MemberBase>();
-
-            private static readonly Dictionary<Type, ClassMember> CachedClassMembers = new Dictionary<Type, ClassMember>();
-
-            private ClassMember(IEnumerable<FieldInfo> fieldInfos, IEnumerable<PropertyInfo> propertyInfos)
-            {
-                foreach (var fieldInfo in fieldInfos)
-                {
-                    members.Add(new FieldMember(fieldInfo));
-                }
-
-                foreach (var propertyInfo in propertyInfos)
-                {
-                    members.Add(new PropertyMember(propertyInfo));
-                }
-            }
-
-            private static T[] GetMembersFromType<T>(Type type, Func<Type, T[]> getMembers)
-            {
-                var members = getMembers(type);
-                var baseType = type.BaseType;
-
-                while (baseType != null)
-                {
-                    var parentMembers = getMembers(baseType);
-                    members = members.Union(parentMembers).ToArray();
-                    baseType = baseType.BaseType;
-                }
-
-                return members;
-            }
-
-            public static ClassMember GetClassMember(Type type)
-            {
-                if (CachedClassMembers.TryGetValue(type, out ClassMember classMember))
-                {
-                    return classMember;
-                }
-
-                const BindingFlags flags = ReflectionHelper.AllBindingFlags & ~BindingFlags.Static;
-                var fields = GetMembersFromType(type, t => t.GetFields(flags));
-                var properties = GetMembersFromType(type, t => t.GetProperties(flags));
-
-                classMember = new ClassMember(fields, properties);
-                CachedClassMembers[type] = classMember;
-                return classMember;
-            }
-        }
-
-        private abstract class MemberBase
-        {
-            public abstract object GetValue(object obj);
-            public abstract void SetValue(object obj, object value);
-            public bool HasGetMethod { get; protected set; }
-            public Type MemberType { get; protected set; }
-            public Type EnumeratedType { get; protected set; }
-            public bool IsUnityObject { get; protected set; }
-            public bool IsClass { get; protected set; }
-            public bool IsEnumerableOfUnityObjects { get; protected set; }
-            public bool IsEnumeratedClass { get; protected set; }
-        }
-
-        private class FieldMember : MemberBase
-        {
-            private readonly FieldInfo fieldInfo;
-
-            public FieldMember(FieldInfo fieldInfo)
-            {
-                this.fieldInfo = fieldInfo;
-                MemberType = fieldInfo.FieldType;
-                IsUnityObject = MemberType.IsSameOrSubclass(typeof(Object));
-                IsClass = MemberType.IsClass;
-                HasGetMethod = true;
-                EnumeratedType = MemberType.GetEnumeratedType();
-                IsEnumerableOfUnityObjects = EnumeratedType?.IsSameOrSubclass(typeof(Object)) == true;
-                IsEnumeratedClass = EnumeratedType?.IsClass == true;
-            }
-
-            public override object GetValue(object obj)
-            {
-                return fieldInfo.GetValue(obj);
-            }
-
-            public override void SetValue(object obj, object value)
-            {
-                fieldInfo.SetValue(obj, value);
-            }
-        }
-
-        private class PropertyMember : MemberBase
-        {
-            private readonly PropertyInfo propertyInfo;
-
-            public PropertyMember(PropertyInfo propertyInfo)
-            {
-                this.propertyInfo = propertyInfo;
-                MemberType = propertyInfo.PropertyType;
-                IsUnityObject = MemberType.IsSameOrSubclass(typeof(Object));
-                IsClass = MemberType.IsClass;
-                HasGetMethod = propertyInfo.GetIndexParameters().Length == 0 && propertyInfo.GetMethod != null;
-                EnumeratedType = MemberType.GetEnumeratedType();
-                IsEnumerableOfUnityObjects = EnumeratedType?.IsSameOrSubclass(typeof(Object)) == true;
-                IsEnumeratedClass = EnumeratedType?.IsClass == true;
-            }
-
-            public override object GetValue(object obj)
-            {
-                return propertyInfo.GetValue(obj);
-            }
-
-            public override void SetValue(object obj, object value)
-            {
-                propertyInfo.SetValue(obj, value);
             }
         }
